Migrate UI JavaScript to TypeScript and simplify home content
