+++ ui/index.html
@@ -7,6 +7,6 @@
   </head>
   <body id="app-body" class="bg-slate-50">
     <div id="root"></div>
-    <script type="module" src="/src/main.jsx"></script>
+    <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
