+++ ui/src/__tests__/developer-mode-toggle.test.tsx
