+++ ui/src/components/developer-mode-toggle.tsx
