+++ ui/vite.config.ts
@@ -4,7 +4,7 @@ import react from '@vitejs/plugin-react'
 export default defineConfig({
   plugins: [react()],
   test: {
-    environment: "jsdom",
-    setupFiles: "./src/test-setup.js"
+    environment: 'jsdom',
+    setupFiles: './src/test-setup.ts'
   }
 })
